refactor(app): move the skills browser state into ainb-app
SkillsViewState, the provider and tab enums and the query filters move to
ainb_app::components::skills. The renderer stays in ainb-core and
re-exports them; helpers it calls become pub.
